[#61718846] Correct team redirect tests
Correct tests so that if more users are created in the test database,
we're not just testing for 'teams/1'. This failed when merged into the
main branch.

Instead, in the controller spec, we test that we're redirected to the
path of the `@team` instance variable.

In the team steps, we test that we're redirected to a URL that matches
the `team/[number]` pattern.

TODO: In future, the Cucumber test should more closely match what the
user should see on the page, rather than the URL.